chore: refactor loose files into utils folders and respective node folders
