+++ app/agent/nodes/resolve_integrations/__init__.py
@@ -1,0 +1,5 @@
+"""Resolve integrations node — integration discovery for investigations."""
+
+from app.agent.nodes.resolve_integrations.node import resolve_integrations
+
+__all__ = ["resolve_integrations"]
