+++ app/agent/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Shared helper modules for agent orchestration."""
