+++ tests/benchmarks/cloudopsbench/tests/test_bench_agent.py
@@ -218,7 +218,7 @@ def test_pure_baseline_agent_overrides_system_prompt() -> None:
     opensre's. Pin that the override returns a string distinct from the
     default ``build_system_prompt`` so a future hook removal can't
     silently turn this back into a flavor of BaselineLLMAloneAgent."""
-    from app.agent.prompt import build_system_prompt
+    from app.agent.nodes.investigate.prompt import build_system_prompt
     from tests.benchmarks.cloudopsbench.bench_agent import PureBaselineAgent
 
     agent = PureBaselineAgent()
